requirements.txt
Pillow==9.1.0
requests==2.27.1
streamlit==1.8.1
